+++ tests/libtest/lib530.c
@@ -162,7 +162,7 @@ static int curlTimerCallback(CURLM *multi, long timeout_ms, void *userp)
   }
   if(timeout_ms != -1) {
     *timeout = tutil_tvnow();
-    timeout->tv_usec += timeout_ms * 1000;
+    timeout->tv_usec += (int)timeout_ms * 1000;
   }
   else {
     timeout->tv_sec = -1;
