+++ tests/libtest/lib582.c
@@ -136,7 +136,7 @@ static int curlTimerCallback(CURLM *multi, long timeout_ms, void *userp)
   (void)multi; /* unused */
   if(timeout_ms != -1) {
     *timeout = tutil_tvnow();
-    timeout->tv_usec += timeout_ms * 1000;
+    timeout->tv_usec += (int)timeout_ms * 1000;
   }
   else {
     timeout->tv_sec = -1;
